Remove gentoo-sources from system set
We used to compile the kernel on each host, using a custom config
(a generic one tuned for low-latency, then modified for our PCs at
rastapank, matzore PCs, radio best etc) but as the number of hosts
using our distro grows I was looking for ways to either use
gentoo-kernel-bin with a few tweaks or gentoo-kernel as a binary
package, with a custom savedconfig but with enough hw support to
be able to work on most recent systems (including laptops and
even some server / heavy workstations etc so that I can use this
kernel in my dev environments as well).

I ended up building a custom gentoo-kernel which is distributed
as a binary package and installed using installkernel-systemd
(which was another thing I wanted to do, since I installed them
manualy on each update). The thing is that unless I mask newer
versions the moment an update happens if I don't update the
package on binhost it'll recompile localy, which is not a problem
in general since it'll still use the savedconfig from the previous
installation. However it becomes a problem when installing the
system for the first time where no savedconfig exists from previous
installations. In that case it makes more sense to just install
gentoo-kernel-bin which is why I didn't add gentoo-kernel in the
system set to allow for more flexibility on the user's side.
The user can obviously also install gentoo-sources and go for a
fully custom local build.

Note that recent kernels allow us to chose preemtion policy at
runtime, so in most cases we can stick to the upstream kernel.
It's still a mess with ancient drivers and in some cases
suboptimal for my taste but it should work, just make sure
you use the following options:

threadirqs mitigations=off preempt=full